- criação do metodo delete para super admin - refatoração de trechos

+++ app/Http/Controllers/VehicleController.php
@@ -104,11 +104,9 @@ class VehicleController extends Controller
     }
     public function destroy(string $id)
     {
-        // Obtém o usuário autenticado
-        $user = auth()->user();
-
+       
         // Verifica se o usuário tem permissão para deletar (usertype 2 ou 3)
-        if ($user->usertype == 2 || $user->usertype == 3) {
+        if (Auth::user()->usertype >=2) {   
             try {
                 $vehicle = Vehicle::findOrFail($id);
                 if ($vehicle->return_status != 'Em análise') {
@@ -116,6 +114,28 @@ class VehicleController extends Controller
                     return redirect(route('dashboard'))
                         ->with('fail', 'Consultas completas não podem ser excluídas.');
                 }
+                Vehicle::destroy($id);
+                return redirect(route('dashboard'))
+                    ->with('success', 'Registro deletado com sucesso');
+            } catch (ValidationException $e) {
+                // Armazena os dados na sessão para depuração
+                return redirect(route('dashboard'))
+                    ->with('fail', 'Falha na validação dos dados: ' . $e->getMessage());
+            } catch (\Exception $e) {
+                // Armazena os dados na sessão para depuração
+                return redirect(route('dashboard'))
+                    ->with('fail', 'Falha no registro: ' . $e->getMessage());
+            }
+        } else {
+            return redirect(route('dashboard'))->with('fail', 'Você não tem permissão para deletar este registro.');
+        }
+    }
+    public function delete(string $id)
+    {
+
+        if ( Auth::user()->usertype == 3) {
+            try {
+
                 Vehicle::destroy($id);
                 return redirect(route('dashboard'))
                     ->with('success', 'Registro deletado com sucesso');
